refactor: import StateProvider explicitly
