+++ src/blocks/dataTable/edit.tsx
@@ -6,7 +6,8 @@ import { PanelBody, PanelRow, ToolbarButton, ToolbarGroup } from '@wordpress/com
 import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Button, Group, SetupEditorEnv, SourceSelect, Stack, StateProvider, Switch, Text, TextInput, useGlobalState } from '../../components'
+import { Button, Group, SetupEditorEnv, SourceSelect, Stack, Switch, Text, TextInput } from '../../components'
+import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import json from './block.json'
 import { Attributes } from './index'
 import { GlobalState, storeCreator } from './state'
