+++ src/blocks/dataTable/view.tsx
@@ -4,7 +4,8 @@ import cloneDeep from 'lodash.clonedeep'
 import type { MRT_ColumnFiltersState as ColumnFiltersState, MRT_SortingState as SortingState } from 'mantine-react-table'
 import { MantineReactTable, useMantineReactTable } from 'mantine-react-table'
 import { useDeepCompareEffect } from 'react-use'
-import { Box, OptionalPortal, useGlobalState } from '../../components'
+import { Box, OptionalPortal } from '../../components'
+import { useGlobalState } from '../../components/StateProvider'
 import { Z_INDEX_ABOVE_ADMIN } from '../../utils'
 import { GlobalState } from './state'
 import { isValueValid } from './utils'
