+++ src/blocks/export/edit.tsx
@@ -5,7 +5,8 @@ import { PanelBody, PanelRow } from '@wordpress/components'
 import { select } from '@wordpress/data'
 import { useEffect } from '@wordpress/element'
 import { disablePostExport, enablePostExport } from '../../ApiServer'
-import { SetupEditorEnv, StateProvider, TextInput, useGlobalState } from '../../components'
+import { SetupEditorEnv, TextInput } from '../../components'
+import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import config from './block.json'
 import { Attributes } from './index'
 import { GlobalState, storeCreator } from './state'
