+++ src/blocks/export/hydration.tsx
@@ -1,7 +1,8 @@
 import { InseriRoot } from '@inseri/lighthouse'
 import domReady from '@wordpress/dom-ready'
 import { createRoot } from '@wordpress/element'
-import { InseriThemeProvider, StateProvider } from '../../components'
+import { InseriThemeProvider } from '../../components'
+import { StateProvider } from '../../components/StateProvider'
 import { generateQuerySelector } from '../../utils'
 import config from './block.json'
 import { storeCreator } from './state'
