+++ src/blocks/iiifViewer/edit.tsx
@@ -6,7 +6,8 @@ import { PanelBody, PanelRow, TextControl, ToggleControl, ToolbarButton, Toolbar
 import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, Group, SetupEditorEnv, SourceSelect, StateProvider, Text, useGlobalState } from '../../components'
+import { Box, Group, SetupEditorEnv, SourceSelect, Text } from '../../components'
+import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import json from './block.json'
 import { Attributes } from './index'
 import { GlobalState, storeCreator } from './state'
