+++ src/blocks/localFileImport/view.tsx
@@ -2,7 +2,8 @@ import { usePublish } from '@inseri/lighthouse'
 import { Dropzone } from '@mantine/dropzone'
 import { IconCheck, IconDownload, IconX } from '@tabler/icons-react'
 import { useEffect, useRef } from '@wordpress/element'
-import { ActionIcon, Button, Group, Stack, Text, createStyles, useGlobalState, useMantineTheme } from '../../components'
+import { ActionIcon, Button, Group, Stack, Text, createStyles, useMantineTheme } from '../../components'
+import { useGlobalState } from '../../components/StateProvider'
 import { guessContentTypeByExtension, handleBody } from '../../utils'
 import { GlobalState } from './state'
