+++ src/blocks/mediaCollection/edit.tsx
@@ -7,7 +7,8 @@ import { useDispatch as useWpDispatch } from '@wordpress/data'
 import { useEffect } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { SetupEditorEnv, StateProvider, useGlobalState } from '../../components'
+import { SetupEditorEnv } from '../../components'
+import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import { HidingWrapper } from '../../utils'
 import config from './block.json'
 import { Attributes } from './index'
