+++ src/blocks/plotly/edit.tsx
@@ -7,7 +7,8 @@ import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
 import { useMap } from 'react-use'
-import { Accordion, ActionIcon, Box, Button, Group, Select, SetupEditorEnv, SourceSelect, Stack, StateProvider, Text, useGlobalState } from '../../components'
+import { Accordion, ActionIcon, Box, Button, Group, Select, SetupEditorEnv, SourceSelect, Stack, Text } from '../../components'
+import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import json from './block.json'
 import { Attributes } from './index'
 import { GlobalState, storeCreator } from './state'
