+++ src/blocks/plotly/view.tsx
@@ -2,7 +2,8 @@ import { Nucleus, usePublish, useRestorableState, useWatch } from '@inseri/light
 import { useState } from '@wordpress/element'
 import cloneDeep from 'lodash.clonedeep'
 import Plot from 'react-plotly.js'
-import { Box, useGlobalState } from '../../components'
+import { Box } from '../../components'
+import { useGlobalState } from '../../components/StateProvider'
 import { GlobalState } from './state'
 import { useDeepCompareEffect } from 'react-use'
