+++ src/blocks/python/view.tsx
@@ -1,7 +1,8 @@
 import { Nucleus, usePublish, useRestorableState, useWatch } from '@inseri/lighthouse'
 import { usePrevious } from '@mantine/hooks'
 import { useEffect } from '@wordpress/element'
-import { Box, CodeEditor, Group, useGlobalState } from '../../components'
+import { Box, CodeEditor, Group } from '../../components'
+import { useGlobalState } from '../../components/StateProvider'
 import { TopBar } from '../../components/TopBar'
 import { Attributes } from './index'
 import { GlobalState } from './state'
