+++ src/blocks/textEditor/view.tsx
@@ -3,7 +3,8 @@ import { useDebouncedValue } from '@mantine/hooks'
 import { IconEye, IconPencil } from '@tabler/icons-react'
 import { useEffect, useMemo, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
-import { Box, Button, CodeEditor, Group, Text, Tooltip, useGlobalState } from '../../components'
+import { Box, Button, CodeEditor, Group, Text, Tooltip } from '../../components'
+import { useGlobalState } from '../../components/StateProvider'
 import { formatCode, getBodyTypeByContenType, isBeautifyType } from '../../utils'
 import { GlobalState } from './state'
