+++ src/blocks/textViewer/edit.tsx
@@ -6,7 +6,8 @@ import { PanelBody, PanelRow, ResizableBox, TextControl, ToolbarButton, ToolbarG
 import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, Group, SetupEditorEnv, SourceSelect, StateProvider, Text, useGlobalState } from '../../components'
+import { Box, Group, SetupEditorEnv, SourceSelect, Text } from '../../components'
+import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import { TEXTUAL_CONTENT_TYPES } from '../../utils'
 import json from './block.json'
 import { Attributes } from './index'
