+++ src/blocks/textViewer/view.tsx
@@ -2,7 +2,8 @@ import { Nucleus, useWatch } from '@inseri/lighthouse'
 import { IconEye } from '@tabler/icons-react'
 import { __ } from '@wordpress/i18n'
 import stringify from 'json-stable-stringify'
-import { Box, CodeEditor, Group, Text, Tooltip, useGlobalState } from '../../components'
+import { Box, CodeEditor, Group, Text, Tooltip } from '../../components'
+import { useGlobalState } from '../../components/StateProvider'
 import { TEXTUAL_CONTENT_TYPES, getBodyTypeByContenType } from '../../utils'
 import { GlobalState } from './state'
