+++ src/blocks/webApi/view.tsx
@@ -1,7 +1,8 @@
 import { Nucleus, usePublish, useWatch } from '@inseri/lighthouse'
 import { useEffect } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
-import { Box, Button, Text, useGlobalState } from '../../components'
+import { Box, Button, Text } from '../../components'
+import { useGlobalState } from '../../components/StateProvider'
 import { DatasourceState } from './AdminState'
 
 interface ViewProps {
