+++ src/components/DetailViewBody.tsx
@@ -2,7 +2,8 @@ import { useDebouncedValue } from '@mantine/hooks'
 import { IconCircleOff } from '@tabler/icons-react'
 import { useEffect } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
-import { Accordion, Alert, Box, Button, CodeEditor, createStyles, Group, SegmentedControl, Tabs, Text, useGlobalState } from '.'
+import { Accordion, Alert, Box, Button, CodeEditor, createStyles, Group, SegmentedControl, Tabs, Text } from '.'
+import { useGlobalState } from './StateProvider'
 import { formatCode, isBeautifyType, isFormType } from '../utils'
 import { ParamsTable } from './ParamsTable'
 import { DatasourceState } from '../blocks/webApi/AdminState'
