+++ src/components/index.tsx
@@ -4,4 +4,3 @@ export * from './SourceSelect'
 export * from './CodeEditor'
 export * from './InseriThemeProvider'
 export * from './SetupEditorEnv'
-export * from './StateProvider'
